requirementsnew.txt
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo==0.2.1
blis==0.7.9
catalogue==1.0.2
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
cryptography==41.0.1
cymem==2.0.7
Django==2.2.10
django-crispy-forms==1.7.2
djangorestframework==3.14.0
docutils==0.16
docx2txt==0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
gunicorn==20.1.0
idna==3.4
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
langcodes==3.3.0
MarkupSafe==2.1.2
murmurhash==1.0.9
nltk==3.8.1
numpy==1.24.3
packaging==23.1
pandas==2.0.2
pathy==0.10.1
pdfminer.six==20221105
pkgutil_resolve_name==1.3.10
plac==1.1.3
preshed==3.0.8
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.8
pyresparser==1.0.6
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
regex==2023.5.5
requests==2.31.0
rsa==4.7.2
s3transfer==0.6.0
six==1.16.0
smart-open==6.3.0
sortedcontainers==2.4.0
spacy==2.3.5
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.4
srsly==1.0.6
thinc==7.4.6
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.2
wasabi==0.10.1
xlwt==1.3.0
zipp==3.15.0
